+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased
+- Add `only-with-attachment` filter on get comments
 - Retry when putting comments
 - Add ability to get email by id 
 - Add ability to upload attachment content for comments